Struct comlib_math::Quot

pub struct Quot<T: Integer> { /* fields omitted */ }

A quotient. Represents a rational number as numerator/denominator.

The denominator is stored as non-zeroable type, so this type benefits from the niche optimization.

Implementations

impl<T: Integer> Quot<T>

pub fn new(numerator: T, denominator: T) -> Option<Self>

Constructs a new Quot from a numerator and a denominator.

Return None if the denominator is 0.

pub fn new_nonzero(numerator: T, denominator: T::NonZero) -> Self

Constructs a new Quot from a numerator and a non-zero denominator.

pub fn zero() -> Self

Gets the 0-value.

pub fn numerator(self) -> T

Gets the numerator of the quotient.

pub fn denominator(self) -> T

Gets the denominator of the quotient.

pub fn trunc(self) -> T

Truncates the quotient.

This gives out the integral part of the quotient.

pub fn abs(self) -> Self where
T: Signed,

Computes the absolute value of the quotient.