Crate comlib_math

Comlib Math Utilities

This library contains some commonly used math utilities.

Content

Greatest common divisor

Modular integers

Sieve of Eratosthenes

Primality test

Factorization

Modular exponentiation

Finding next permutation of a list

Iterating over subsets

Still missing

Fourier Transform, both number theoretic and complex

Re-exports

pub use subsets::subsets;

Modules

subsets
Module containing functions and structs for iterating subsets.

Structs

Mod1e9p7
Commonly used modulus 10⁹ + 7.

ModInt
Integer type for which all computations are executed in modulo given by M::modulus.

PrimeSieve
Sieve of Eratosthenes.

RuntimeModulus
Modulus whose value can be selected at runtime.

RuntimePrimeModulus
Modulus whose value can be selected at runtime.

Traits

Integer
Trait implemented by integers.

InvertibleModulus
Marker trait for indicating that modular numbers in this modulo can be inverted.

Modulus
Modulus contains the modulus and the actual storage type of ModInt.

Functions

factorize
Factorizes the given integer into its prime factors.

gcd
Computes the greatest common divisor of the given numbers.

is_prime
Checks whether a given number is a prime.

mod_pow
Raises base to given exponent in the given modulus.

next_permutation
Computes lexicographically next smallest permutation.