Trait comlib_math::Modulus

pub trait Modulus {
    type Base: Copy + Display + Debug + Add<Output = Self::Base> + Sub<Output = Self::Base> + Mul<Output = Self::Base> + Div<Output = Self::Base> + Rem<Output = Self::Base> + AddAssign + SubAssign + MulAssign + DivAssign + RemAssign + Eq + Ord + Default + From<u8>;
    fn modulus(self) -> Self::Base;

    fn inverse_power(self) -> usize;
}

Modulus contains the modulus and the actual storage type of ModInt.

Associated Types

type Base: Copy + Display + Debug + Add<Output = Self::Base> + Sub<Output = Self::Base> + Mul<Output = Self::Base> + Div<Output = Self::Base> + Rem<Output = Self::Base> + AddAssign + SubAssign + MulAssign + DivAssign + RemAssign + Eq + Ord + Default + From<u8>

Type for holding values mod modulus().

Must be large enough to contain the square of modulus() - 1, that is the maximum value which can be encountered during multiplication.

Loading content...
Required methods

fn modulus(self) -> Self::Base

Modulus in which computations should be done.

fn inverse_power(self) -> usize

The power a value needs to be raised to in order to get its modular inverse.

If the modulus is a prime, then this is modulus() - 2 by Fermat's little theorem. For composite numbers this must be φ(modulus()) - 1 where φ is the Euler's totient function which gives the number of integers smaller than modulus() which are coprime with modulus().

Loading content...